Switched Random example from int to size_t

+++ examples/tools/Random.cc
@@ -16,10 +16,10 @@ int main()
 
   std::cout << "Digits in random orders:" << std::endl;
 
-  int num_samples = 10;
-  for (int s  = 0; s < num_samples; s++) {
-    emp::vector<int> permut = emp::GetPermutation(random, 10);
-    for (int i = 0; i < 10; i++) {
+  size_t num_samples = 10;
+  for (size_t s  = 0; s < num_samples; s++) {
+    emp::vector<size_t> permut = emp::GetPermutation(random, 10);
+    for (size_t i = 0; i < 10; i++) {
       std::cout << permut[i];
     }
     std::cout << std::endl;
